вынес аффекты и функции манипуляции ими в отдельный файл. перетащил файл с логикой поведения мобов - там только агро - в fightsystem и причесал хидера
