+++ src/magic.cpp
@@ -14,6 +14,7 @@
 
 #include "magic.h"
 
+#include "core/affect_data.h"
 #include "action.targeting.hpp"
 #include "chars/world.characters.hpp"
 #include "world.objects.hpp"
@@ -749,193 +750,6 @@ void show_spell_off(int aff, CHAR_DATA * ch)
 	}
 }
 
-void mobile_affect_update(void)
-{
-	character_list.foreach_on_copy([](const CHAR_DATA::shared_ptr& i)
-	{
-		int was_charmed = FALSE, charmed_msg = FALSE;
-		bool was_purged = false;
-
-		if (IS_NPC(i))
-		{
-			auto next_affect_i = i->affected.begin();
-			for (auto affect_i = next_affect_i; affect_i != i->affected.end(); affect_i = next_affect_i)
-			{
-				++next_affect_i;
-				const auto& affect = *affect_i;
-
-				if (affect->duration >= 1)
-				{
-					if (IS_SET(affect->battleflag, AF_SAME_TIME) && (!i->get_fighting() || affect->location == APPLY_POISON))
-					{
-						// здесь плеера могут спуржить
-						if (processPoisonDamage(i.get(), affect) == -1)
-						{
-							was_purged = true;
-
-							break;
-						}
-					}
-
-					affect->duration--;
-					if (affect->type == SPELL_CHARM && !charmed_msg && affect->duration <= 1)
-					{
-						act("$n начал$g растерянно оглядываться по сторонам.", FALSE, i.get(), 0, 0, TO_ROOM | TO_ARENA_LISTEN);
-						charmed_msg = TRUE;
-					}
-				}
-				else if (affect->duration == -1)
-				{
-					affect->duration = -1;	// GODS - unlimited
-				}
-				else
-				{
-					if (affect->type > 0
-						&& affect->type <= MAX_SPELLS)
-					{
-						if (next_affect_i == i->affected.end()
-							|| (*next_affect_i)->type != affect->type
-							|| (*next_affect_i)->duration > 0)
-						{
-							if (affect->type > 0
-								&& affect->type <= SPELLS_COUNT
-								&& *spell_wear_off_msg[affect->type])
-							{
-								show_spell_off(affect->type, i.get());
-								if (affect->type == SPELL_CHARM
-									|| affect->bitvector == to_underlying(EAffectFlag::AFF_CHARM))
-								{
-									was_charmed = TRUE;
-								}
-							}
-						}
-					}
-
-					i->affect_remove(affect_i);
-				}
-			}
-		}
-
-		if (!was_purged)
-		{
-			affect_total(i.get());
-
-			decltype(i->timed) timed_next;
-			for (auto timed = i->timed; timed; timed = timed_next)
-			{
-				timed_next = timed->next;
-				if (timed->time >= 1)
-				{
-					timed->time--;
-				}
-				else
-				{
-					timed_from_char(i.get(), timed);
-				}
-			}
-
-			for (auto timed = i->timed_feat; timed; timed = timed_next)
-			{
-				timed_next = timed->next;
-				if (timed->time >= 1)
-				{
-					timed->time--;
-				}
-				else
-				{
-					timed_feat_from_char(i.get(), timed);
-				}
-			}
-
-			if (DeathTrap::check_death_trap(i.get()))
-			{
-				return;
-			}
-
-			if (was_charmed)
-			{
-				stop_follower(i.get(), SF_CHARMLOST);
-			}
-		}
-	});
-}
-
-void player_affect_update(void)
-{
-	character_list.foreach_on_copy([](const CHAR_DATA::shared_ptr& i)
-	{
-		// на всякий случай проверка на пурж, в целом пурж чармисов внутри
-		// такого цикла сейчас выглядит безопасным, чармисы если и есть, то они
-		// добавлялись в чар-лист в начало списка и идут до самого чара
-		if (i->purged()
-			|| IS_NPC(i)
-			|| DeathTrap::tunnel_damage(i.get()))
-		{
-			return;
-		}
-
-		pulse_affect_update(i.get());
-
-		bool was_purged = false;
-		auto next_affect_i = i->affected.begin();
-		for (auto affect_i = next_affect_i; affect_i != i->affected.end(); affect_i = next_affect_i)
-		{
-			++next_affect_i;
-			const auto& affect = *affect_i;
-
-			if (affect->duration >= 1)
-			{
-				if (IS_SET(affect->battleflag, AF_SAME_TIME) && !i->get_fighting())
-				{
-					// здесь плеера могут спуржить
-					if (processPoisonDamage(i.get(), affect) == -1)
-					{
-						was_purged = true;
-						break;
-					}
-				}
-				affect->duration--;
-			}
-			else if (affect->duration != -1)
-			{
-				if ((affect->type > 0) && (affect->type <= MAX_SPELLS))
-				{
-					if (next_affect_i == i->affected.end()
-						|| (*next_affect_i)->type != affect->type
-						|| (*next_affect_i)->duration > 0)
-					{
-						if (affect->type > 0
-							&& affect->type <= SPELLS_COUNT
-							&& *spell_wear_off_msg[affect->type])
-						{
-							//чтобы не выдавалось, "что теперь вы можете сражаться",
-							//хотя на самом деле не можете :)
-							if (!(affect->type == SPELL_MAGICBATTLE
-								&& AFF_FLAGGED(i, EAffectFlag::AFF_STOPFIGHT)))
-							{
-								if (!(affect->type == SPELL_BATTLE
-									&& AFF_FLAGGED(i, EAffectFlag::AFF_MAGICSTOPFIGHT)))
-								{
-									show_spell_off(affect->type, i.get());
-								}
-							}
-						}
-					}
-				}
-
-				i->affect_remove(affect_i);
-			}
-		}
-
-		if (!was_purged)
-		{
-			MemQ_slots(i.get());	// сколько каких слотов занято (с коррекцией)
-
-			affect_total(i.get());
-		}
-	});
-}
-
 // зависимость длительности закла от скила магии
 float func_koef_duration(int spellnum, int percent)
 {
@@ -981,115 +795,7 @@ float func_koef_modif(int spellnum, int percent)
 	return 0;
 }
 
-// This file update battle affects only
-void battle_affect_update(CHAR_DATA * ch)
-{
-	auto next_affect_i = ch->affected.begin();
-	for (auto affect_i = next_affect_i; affect_i != ch->affected.end(); affect_i = next_affect_i)
-	{
-		++next_affect_i;
-		const auto& affect = *affect_i;
 
-		if (!IS_SET(affect->battleflag, AF_BATTLEDEC) && !IS_SET(affect->battleflag, AF_SAME_TIME))
-			continue;
-
-		if (IS_NPC(ch) && affect->location == APPLY_POISON)
-			continue;
-
-		if (affect->duration >= 1) {
-			if (IS_SET(affect->battleflag, AF_SAME_TIME)) {
-			    if (processPoisonDamage(ch, affect) == -1) // жертва умерла
-					return;
-				affect->duration--;
-			}
-			else {
-				if (IS_NPC(ch))
-					affect->duration--;
-				else
-					affect->duration -= MIN(affect->duration, SECS_PER_MUD_HOUR / SECS_PER_PLAYER_AFFECT);
-			}
-		}
-		else if (affect->duration != -1) {
-			if (affect->type > 0 && affect->type <= MAX_SPELLS) {
-				if (next_affect_i == ch->affected.end()
-					|| (*next_affect_i)->type != affect->type
-					|| (*next_affect_i)->duration > 0)
-				{
-					if (affect->type > 0 && affect->type <= SPELLS_COUNT && *spell_wear_off_msg[affect->type])
-						show_spell_off(affect->type, ch);
-				}
-			}
-
-			ch->affect_remove(affect_i);
-		}
-	}
-
-	affect_total(ch);
-}
-
-// This file update pulse affects only
-void pulse_affect_update(CHAR_DATA * ch)
-{
-	bool pulse_aff = FALSE;
-
-	if (ch->get_fighting())
-	{
-		return;
-	}
-
-	auto next_affect_i = ch->affected.begin();
-	for (auto affect_i = next_affect_i; affect_i != ch->affected.end(); affect_i = next_affect_i)
-	{
-		++next_affect_i;
-		const auto& affect = *affect_i;
-
-		if (!IS_SET(affect->battleflag, AF_PULSEDEC))
-		{
-			continue;
-		}
-
-		pulse_aff = TRUE;
-		if (affect->duration >= 1)
-		{
-			if (IS_NPC(ch))
-			{
-				affect->duration--;
-			}
-			else
-			{
-				affect->duration -= MIN(affect->duration, SECS_PER_PLAYER_AFFECT * PASSES_PER_SEC);
-			}
-		}
-		else if (affect->duration == -1)	// No action //
-		{
-			affect->duration = -1;	// GODs only! unlimited //
-		}
-		else
-		{
-			if ((affect->type > 0) && (affect->type <= MAX_SPELLS))
-			{
-				if (next_affect_i == ch->affected.end()
-					|| (*next_affect_i)->type != affect->type
-					|| (*next_affect_i)->duration > 0)
-				{
-					if (affect->type > 0
-						&& affect->type <= SPELLS_COUNT
-						&& *spell_wear_off_msg[affect->type])
-					{
-						show_spell_off(affect->type, ch);
-					}
-				}
-			}
-
-			ch->affect_remove(affect_i);
-		}
-	}
-
-	if (pulse_aff)
-	{
-		affect_total(ch);
-	}
-}
 
 /*
  *  mag_materials:
@@ -1575,19 +1281,6 @@ int check_recipe_values(CHAR_DATA * ch, int spellnum, int spelltype, int showrec
 
 	return (TRUE);
 }
-
-//******************************************************************************
-//******************************************************************************
-//******************************************************************************
-
-/*
- * Every spell that does damage comes through here.  This calculates the
- * amount of damage, adds in any modifiers, determines what the saves are,
- * tests for save and calls damage().
- *
- * -1 = dead, otherwise the amount of damage done.
- */
-
 
 int magic_skill_damage_calc(CHAR_DATA * ch, CHAR_DATA * victim, int spellnum, int dam) {
 	if (IS_NPC(ch)) {
